+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Updated the **upload** command to support indicator fields.
 * The **validate** and **update-release-notes** commands will now check changed files against `demisto/master` if it is configured locally.
 * Fixed an issue where **validate** would incorrectly identify files as renamed.
+* Fixed an issue where **validate** failed when comparing branch against commit hash.
 
 # 1.3.0
 * Do not collect optional dependencies on indicator types reputation commands.